move all non-route content into lib/
