+++ frontend/src/lib/components/avatar.svelte
