+++ frontend/src/lib/components/back_to_top_button.svelte
