+++ frontend/src/lib/components/collapse.svelte
