+++ frontend/src/lib/components/details/person.svelte
