+++ frontend/src/lib/components/details/person_media.svelte
