+++ frontend/src/lib/components/details/read_more.svelte
