+++ frontend/src/lib/components/details/recommendations.svelte
