+++ frontend/src/lib/components/details/seasons.svelte
