+++ frontend/src/lib/components/error.svelte
