+++ frontend/src/lib/components/error_buttons.svelte
