+++ frontend/src/lib/components/filters.svelte
