+++ frontend/src/lib/components/footer.svelte
