+++ frontend/src/lib/components/loading/spinner.svelte
