+++ frontend/src/lib/components/navbar/navbar.svelte
