+++ frontend/src/lib/components/navbar/phone_navbar.svelte
