+++ frontend/src/lib/components/no_results.svelte
