+++ frontend/src/lib/components/theme_selector.svelte
