+++ frontend/src/lib/variables.js
