+++ frontend/src/routes/+layout.svelte
@@ -1,12 +1,12 @@
 <script>
   import "../app.css"
-  import Navbar from "../components/navbar/navbar.svelte"
-  import PhoneNavbar from "../components/navbar/phone_navbar.svelte"
-  import Footer from "../components/footer.svelte"
-  import BackToTopButton from "../components/back_to_top_button.svelte"
-  import CookieDisclaimer from "../components/cookie_disclaimer.svelte"
+  import Navbar from "$lib/components/navbar/navbar.svelte"
+  import PhoneNavbar from "$lib/components/navbar/phone_navbar.svelte"
+  import Footer from "$lib/components/footer.svelte"
+  import BackToTopButton from "$lib/components/back_to_top_button.svelte"
+  import CookieDisclaimer from "$lib/components/cookie_disclaimer.svelte"
   import MediaQuery from "svelte-media-query"
-  import CountryLocator from "../components/country_locator.svelte"
+  import CountryLocator from "$lib/components/country_locator.svelte"
 </script>
 
 <div class="flex flex-col h-screen">
