+++ frontend/src/routes/+page.svelte
@@ -1,18 +1,18 @@
 <script lang="ts">
-  import { calculateAmountOfShownItems } from "../utils"
-  import { GO_API, PYTHON_API } from "../variables.js"
+  import { calculateAmountOfShownItems } from "$lib/utils"
+  import { GO_API, PYTHON_API } from "$lib/variables.js"
   import Select from "svelte-select"
-  import MediaSearch from "../components/media_search.svelte"
-  import Filters from "../components/filters.svelte"
-  import { currentCountry } from "../stores/country.js"
-  import { currentProviders } from "../stores/providers.js"
-  import { currentGenres } from "../stores/genres"
-  import { inputQuery } from "../stores/input"
-  import { chosenTheme } from "../stores/theme.js"
-  import { filters } from "../stores/filters.js"
+  import MediaSearch from "$lib/components/media_search.svelte"
+  import Filters from "$lib/components/filters.svelte"
+  import { currentCountry } from "$lib/stores/country.js"
+  import { currentProviders } from "$lib/stores/providers.js"
+  import { currentGenres } from "$lib/stores/genres"
+  import { inputQuery } from "$lib/stores/input"
+  import { chosenTheme } from "$lib/stores/theme.js"
+  import { filters } from "$lib/stores/filters.js"
   import { onMount } from "svelte"
   import DT from "daisyui/colors/themes.js"
-  import type { Genre, Media, Meilisearch } from "../types"
+  import type { Genre, Media, Meilisearch } from "$lib/types"
 
   const SEARCH_URL = `${PYTHON_API}/search/`
   const GENRE_URL = `${GO_API}/genres/`
